Menu Update - New "Man" sprite, now with animation - New Mothership sprite - Included a bitmap font called "EdoSZ" - Added a new "How To Play" screen - Added a new "Main Menu" screen - Shuffled the load order of states around to accomodate the new states.
Next update may or may not have more art. Implementing a pause menu is major feature which needs to be addressed soon. I also need to finish up the gameplay state by including the timer. I am assuming a timer will be easy to implement, but a pain to test with, so I have been putting it off.

Once those are finished, I hope to create my first REAL test level. At that point, a hard decision will need to be made. Will the levels be handcrafted, or randomly generated using a hopefully fine tuned algorithm.