Have to stop fighting this glitch There is a problem I did not forsee. It seems that for some reason or another, I can only create at max 3  men at the bottom the screen. I can display more, but if I do, then update() no longer gets called.
I'm going to create a new branch, taking what I have learned so far, and rework the entire Man class. My plan of attack is to move all the extra bits into the default "data" object reference that is built into a sprite. If I do that, I should be able to use object batching to make this all work.

Let me be clear, this is a deciding moment the project. If I cannot resolve this glitch, the game is doomed. 3 men are not enough to make this game engaging, and will certainly kill any chance of success this game might have.