Back on Track
Everything is now more or less working like how it was. The man class is now completely removed from the picture. Instead, I am now much more dependent on the default sprite properties and physics properties.

There is a new script called "AlienProperties" which should now be assigned to the "data" property of the sprite class when creating aliens/men. In other languages, AlienProperties would be treated as a struct, and should only hold additional data, not modify it.